+++ src/complex/small_complex.rs
@@ -137,12 +137,12 @@ impl SmallComplex {
     /// let mut c = SmallComplex::from((1.0f32, 3.0f32));
     /// // rotation does not change the precision
     /// unsafe {
-    ///     c.as_nongrowing_mut().mul_i_mut(false);
+    ///     c.as_nonreallocating_mut().mul_i_mut(false);
     /// }
     /// assert_eq!(*c, (-3.0, 1.0));
     /// ```
     #[inline]
-    pub unsafe fn as_nongrowing_mut(&mut self) -> &mut Complex {
+    pub unsafe fn as_nonreallocating_mut(&mut self) -> &mut Complex {
         self.update_d();
         let ptr = (&mut self.re) as *mut _ as *mut _;
         &mut *ptr