+++ src/float/small_float.rs
@@ -129,12 +129,12 @@ impl SmallFloat {
     /// let mut f = SmallFloat::from(1.0f32);
     /// // addition does not change the precision
     /// unsafe {
-    ///     *f.as_nongrowing_mut() += 2.0;
+    ///     *f.as_nonreallocating_mut() += 2.0;
     /// }
     /// assert_eq!(*f, 3.0);
     /// ```
     #[inline]
-    pub unsafe fn as_nongrowing_mut(&mut self) -> &mut Float {
+    pub unsafe fn as_nonreallocating_mut(&mut self) -> &mut Float {
         self.update_d();
         let ptr = (&mut self.inner) as *mut _ as *mut _;
         &mut *ptr